+++ static-site-generator/src/components/ui/layouts/Inner/index.tsx
