+++ static-site-generator/src/components/ui/layouts/PageWrapper/index.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react'
+
+type LayoutPageWrapperProps = {
+  children: ReactNode
+}
+
+export const LayoutPageWrapper = ({ children }: LayoutPageWrapperProps) => {
+  return <div className='LayoutPageWrapper'>{children}</div>
+}
