+++ static-site-generator/src/components/ui/layouts/PageWrapper/style.css
@@ -1,6 +1,6 @@
 @use '~/assets/scss/global/' as *;
 
-.LayoutWrapper {
+.LayoutPageWrapper {
   display: flex;
   flex-direction: column;
   min-height: 100svh;
