--- static-site-generator/src/layouts/Wrapper/index.tsx
@@ -1,9 +1,0 @@
-import { ReactNode } from 'react'
-
-type LayoutWrapperProps = {
-  children: ReactNode
-}
-
-export const LayoutWrapper = ({ children }: LayoutWrapperProps) => {
-  return <div className='LayoutWrapper'>{children}</div>
-}
