+++ static-site-generator/src/stories/components/ui/layouts/Inner/index.stories.tsx
@@ -1,24 +1,24 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
-import { LayoutSection } from '~/components/layout/Section'
+import { LayoutInner } from '~/components/ui/layouts/Inner'
 
 export default {
-  title: 'components/layout/Section',
-  component: LayoutSection,
-} as Meta<typeof LayoutSection>
+  title: 'components/ui/layouts/Inner',
+  component: LayoutInner,
+} as Meta<typeof LayoutInner>
 
-type Story = StoryObj<typeof LayoutSection>
+type Story = StoryObj<typeof LayoutInner>
 
 export const Default: Story = {
   render: (args) => {
     return (
-      <LayoutSection {...args}>
+      <LayoutInner {...args}>
         <>
           コンテンツが入ります
           <br />
           テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト
         </>
-      </LayoutSection>
+      </LayoutInner>
     )
   },
 }
